* commit '6f29e29ed76419e5f3dd0a91ace664a9e5de16ea':
  map to survey input

+++ database/migrations/2016_02_27_205643_create_places_table.php
@@ -24,8 +24,8 @@ class CreatePlacesTable extends Migration
 //                ->onDelete('cascade');
 
 
-            $table->double('lat');
-            $table->double('lng');
+            $table->float('lat');
+            $table->float('lng');
             $table->timestamps();
